only create AttentiveConfig once, separate identify & trigger
