+++ attentive-mobile-sdk/src/main/java/com/attentive/mobilesdk/AttentiveConfig.java
@@ -9,6 +9,7 @@ public class AttentiveConfig {
 
     private final Mode mode;
     private final String domain;
+    private String appUserId;
 
     public AttentiveConfig(String domain, Mode mode) {
         this.domain = domain;
@@ -22,4 +23,8 @@ public class AttentiveConfig {
     public String getDomain() {
         return domain;
     }
+
+    public String getAppUserId() { return appUserId; }
+
+    public void identify(String appUserId) { this.appUserId = appUserId; }
 }
