+++ example/src/main/AndroidManifest.xml
@@ -4,6 +4,7 @@
     package="com.attentive.example">
     <uses-permission android:name="android.permission.INTERNET"/>
     <application
+        android:name=".ExampleApp"
         android:usesCleartextTraffic="false"
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
@@ -15,18 +16,18 @@
         android:theme="@style/Theme.AttentiveExample"
         tools:targetApi="31">
         <activity
-            android:name="com.attentive.example.MainActivity"
+            android:name="com.attentive.example.activities.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name="com.attentive.example.LoadCreativeActivity"
-            android:parentActivityName="com.attentive.example.MainActivity">
+        <activity android:name="com.attentive.example.activities.LoadCreativeActivity"
+            android:parentActivityName="com.attentive.example.activities.MainActivity">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="com.attentive.example.MainActivity" />
+                android:value="com.attentive.example.activities.MainActivity" />
         </activity>
     </application>
