+++ example/src/main/res/layout/activity_load_creative.xml
@@ -5,7 +5,7 @@
     android:id="@+id/loadCreative"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".LoadCreativeActivity">
+    tools:context=".activities.LoadCreativeActivity">
 
     <Button
         android:id="@+id/button"
